Basic isometric view

+++ src/utils/utils.h
@@ -6,7 +6,7 @@
 /*   By: olimarti <olimarti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/07 00:47:16 by olimarti          #+#    #+#             */
-/*   Updated: 2023/03/23 02:20:13 by olimarti         ###   ########.fr       */
+/*   Updated: 2023/03/23 04:26:47 by olimarti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,15 +25,15 @@
 
 typedef struct s_vect2d
 {
-	int	x;
-	int	y;
+	float	x;
+	float	y;
 }	t_vect2d;
 
 typedef struct s_vect3d
 {
-	int	x;
-	int	y;
-	int	z;
+	float	x;
+	float	y;
+	float	z;
 }	t_vect3d;
 
 char		**ft_split(char const *s, char c);
@@ -47,12 +47,12 @@ int			ft_atoi_base(const char *nptr, int *nbr, char *base, int base_len);
 int			ft_atoi_hex(const char *nptr, int *nbr);
 void		ft_putstr_fd(char *str, int fd);
 
-t_vect2d	vect2d(int x, int y);
+t_vect2d	vect2d(float x, float y);
 t_vect2d	vect2d_add(t_vect2d vect1, t_vect2d vect2);
 t_vect2d	vect2d_divide(t_vect2d vect1, t_vect2d vect2);
 t_vect2d	vect2d_multiply(t_vect2d vect1, t_vect2d vect2);
 
-t_vect3d	vect3d(int x, int y, int z);
+t_vect3d	vect3d(float x, float y, float z);
 t_vect3d	vect3d_add(t_vect3d vect1, t_vect3d vect2);
 t_vect3d	vect3d_divide(t_vect3d vect1, t_vect3d vect2);
 t_vect3d	vect3d_multiply(t_vect3d vect1, t_vect3d vect2);
