+++ src/utils/vectors/vect2d/vect2d.c
@@ -6,13 +6,13 @@
 /*   By: olimarti <olimarti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/07 00:51:11 by olimarti          #+#    #+#             */
-/*   Updated: 2023/03/07 23:34:28 by olimarti         ###   ########.fr       */
+/*   Updated: 2023/03/23 04:25:59 by olimarti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../utils.h"
 
-t_vect2d	vect2d(int x, int y)
+t_vect2d	vect2d(float x, float y)
 {
 	t_vect2d	result;
